requirements_test.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --extra test
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
coverage==7.6.1
    # via pytest-cov
fastapi==0.115.0
    # via webhook-receive (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via webhook-receive (pyproject.toml)
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
pytest==8.3.3
    # via
    #   webhook-receive (pyproject.toml)
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-env
pytest-asyncio==0.24.0
    # via webhook-receive (pyproject.toml)
pytest-cov==5.0.0
    # via webhook-receive (pyproject.toml)
pytest-env==1.1.5
    # via webhook-receive (pyproject.toml)
python-dotenv==1.0.1
    # via
    #   webhook-receive (pyproject.toml)
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via webhook-receive (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.38.6
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via requests
uvicorn==0.31.0
    # via webhook-receive (pyproject.toml)
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
